Reshuffle and rename
