+++ lib/tbot/service_workload_identity_x509.go
@@ -27,16 +27,14 @@ import (
 	"time"
 
 	"github.com/gravitational/trace"
-	"google.golang.org/protobuf/types/known/durationpb"
 
 	workloadidentityv1pb "github.com/gravitational/teleport/api/gen/proto/go/teleport/workloadidentity/v1"
 	"github.com/gravitational/teleport/api/utils/retryutils"
 	"github.com/gravitational/teleport/lib/auth/authclient"
-	"github.com/gravitational/teleport/lib/cryptosuites"
 	"github.com/gravitational/teleport/lib/reversetunnelclient"
 	"github.com/gravitational/teleport/lib/tbot/config"
 	"github.com/gravitational/teleport/lib/tbot/identity"
-	"github.com/gravitational/teleport/lib/tbot/spiffe"
+	"github.com/gravitational/teleport/lib/tbot/workloadidentity"
 )
 
 // WorkloadIdentityX509Service is a service that retrieves X.509 certificates
@@ -50,7 +48,7 @@ type WorkloadIdentityX509Service struct {
 	resolver       reversetunnelclient.Resolver
 	// trustBundleCache is the cache of trust bundles. It only needs to be
 	// provided when running in daemon mode.
-	trustBundleCache *spiffe.TrustBundleCache
+	trustBundleCache *workloadidentity.TrustBundleCache
 }
 
 // String returns a human-readable description of the service.
@@ -65,7 +63,7 @@ func (s *WorkloadIdentityX509Service) OneShot(ctx context.Context) error {
 	if err != nil {
 		return trace.Wrap(err, "requesting SVID")
 	}
-	bundleSet, err := spiffe.FetchInitialBundleSet(
+	bundleSet, err := workloadidentity.FetchInitialBundleSet(
 		ctx,
 		s.log,
 		s.botAuthClient.SPIFFEFederationServiceClient(),
@@ -191,7 +189,7 @@ func (s *WorkloadIdentityX509Service) requestSVID(
 	}
 	defer impersonatedClient.Close()
 
-	x509Credentials, privateKey, err := issueX509WorkloadIdentity(
+	x509Credentials, privateKey, err := workloadidentity.IssueX509WorkloadIdentity(
 		ctx,
 		s.log,
 		impersonatedClient,
@@ -219,7 +217,7 @@ func (s *WorkloadIdentityX509Service) requestSVID(
 
 func (s *WorkloadIdentityX509Service) render(
 	ctx context.Context,
-	bundleSet *spiffe.BundleSet,
+	bundleSet *workloadidentity.BundleSet,
 	x509Cred *workloadidentityv1pb.Credential,
 	privateKey crypto.Signer,
 ) error {
@@ -288,118 +286,8 @@ func (s *WorkloadIdentityX509Service) render(
 	s.log.InfoContext(
 		ctx,
 		"Successfully wrote X509 workload identity credential to destination",
-		"workload_identity", workloadIdentityLogValue(x509Cred),
+		"workload_identity", workloadidentity.WorkloadIdentityLogValue(x509Cred),
 		"destination", s.cfg.Destination.String(),
 	)
 	return nil
-}
-
-func workloadIdentityLogValue(credential *workloadidentityv1pb.Credential) slog.Value {
-	return slog.GroupValue(
-		slog.String("name", credential.GetWorkloadIdentityName()),
-		slog.String("revision", credential.GetWorkloadIdentityRevision()),
-		slog.String("spiffe_id", credential.GetSpiffeId()),
-		slog.String("serial_number", credential.GetX509Svid().GetSerialNumber()),
-	)
-}
-
-func workloadIdentitiesLogValue(credentials []*workloadidentityv1pb.Credential) []slog.Value {
-	values := make([]slog.Value, 0, len(credentials))
-	for _, credential := range credentials {
-		values = append(values, workloadIdentityLogValue(credential))
-	}
-	return values
-}
-
-func issueX509WorkloadIdentity(
-	ctx context.Context,
-	log *slog.Logger,
-	clt *authclient.Client,
-	workloadIdentity config.WorkloadIdentitySelector,
-	ttl time.Duration,
-	attest *workloadidentityv1pb.WorkloadAttrs,
-) ([]*workloadidentityv1pb.Credential, crypto.Signer, error) {
-	ctx, span := tracer.Start(
-		ctx,
-		"issueX509WorkloadIdentity",
-	)
-	defer span.End()
-	privateKey, err := cryptosuites.GenerateKey(ctx,
-		cryptosuites.GetCurrentSuiteFromAuthPreference(clt),
-		cryptosuites.BotSVID)
-	if err != nil {
-		return nil, nil, trace.Wrap(err)
-	}
-	pubBytes, err := x509.MarshalPKIXPublicKey(privateKey.Public())
-	if err != nil {
-		return nil, nil, trace.Wrap(err)
-	}
-
-	switch {
-	case workloadIdentity.Name != "":
-		log.DebugContext(
-			ctx,
-			"Requesting issuance of X509 workload identity credential using name of WorkloadIdentity resource",
-			"name", workloadIdentity.Name,
-		)
-		// When using the "name" based selector, we either get a single WIC back,
-		// or an error. We don't need to worry about selecting the right one.
-		res, err := clt.WorkloadIdentityIssuanceClient().IssueWorkloadIdentity(ctx,
-			&workloadidentityv1pb.IssueWorkloadIdentityRequest{
-				Name: workloadIdentity.Name,
-				Credential: &workloadidentityv1pb.IssueWorkloadIdentityRequest_X509SvidParams{
-					X509SvidParams: &workloadidentityv1pb.X509SVIDParams{
-						PublicKey: pubBytes,
-					},
-				},
-				RequestedTtl:  durationpb.New(ttl),
-				WorkloadAttrs: attest,
-			},
-		)
-		if err != nil {
-			return nil, nil, trace.Wrap(err)
-		}
-		log.DebugContext(
-			ctx,
-			"Received X509 workload identity credential",
-			"credential", workloadIdentityLogValue(res.Credential),
-		)
-		return []*workloadidentityv1pb.Credential{res.Credential}, privateKey, nil
-	case len(workloadIdentity.Labels) > 0:
-		labelSelectors := make([]*workloadidentityv1pb.LabelSelector, 0, len(workloadIdentity.Labels))
-		for k, v := range workloadIdentity.Labels {
-			labelSelectors = append(labelSelectors, &workloadidentityv1pb.LabelSelector{
-				Key:    k,
-				Values: v,
-			})
-		}
-		log.DebugContext(
-			ctx,
-			"Requesting issuance of X509 workload identity credentials using labels",
-			"labels", labelSelectors,
-		)
-		res, err := clt.WorkloadIdentityIssuanceClient().IssueWorkloadIdentities(ctx,
-			&workloadidentityv1pb.IssueWorkloadIdentitiesRequest{
-				LabelSelectors: labelSelectors,
-				Credential: &workloadidentityv1pb.IssueWorkloadIdentitiesRequest_X509SvidParams{
-					X509SvidParams: &workloadidentityv1pb.X509SVIDParams{
-						PublicKey: pubBytes,
-					},
-				},
-				RequestedTtl:  durationpb.New(ttl),
-				WorkloadAttrs: attest,
-			},
-		)
-		if err != nil {
-			return nil, nil, trace.Wrap(err)
-		}
-		log.DebugContext(
-			ctx,
-			"Received X509 workload identity credentials",
-			"credentials", workloadIdentitiesLogValue(res.Credentials),
-		)
-		return res.Credentials, privateKey, nil
-	default:
-		return nil, nil, trace.BadParameter("no valid selector configured")
-	}
 }
