+++ lib/tbot/workloadidentity/trust_bundle_cache_test.go
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package spiffe
+package workloadidentity
 
 import (
 	"context"
