+++ lib/tbot/workloadidentity/workloadattest/kubernetes_windows.go
