+++ lib/tbot/workloadidentity/workloadattest/testdata/mountfile/k8s-real-gcp-v1.29.5-gke.1091002
